+++ invokeai/frontend/web/src/features/gallery/store/actions.ts
@@ -5,5 +5,3 @@ export const sentImageToCanvas = createAction('gallery/sentImageToCanvas');
 export const sentImageToImg2Img = createAction('gallery/sentImageToImg2Img');
 
 export const imageDownloaded = createAction('gallery/imageDownloaded');
-
-export const checkAutoAddBoardVisible = createAction('gallery/checkAutoAddBoardVisible');
