+++ invokeai/frontend/web/src/features/gallery/store/gallerySlice.ts
@@ -1,9 +1,7 @@
 import type { PayloadAction } from '@reduxjs/toolkit';
-import { createSlice, isAnyOf } from '@reduxjs/toolkit';
+import { createSlice } from '@reduxjs/toolkit';
 import type { PersistConfig, RootState } from 'app/store/store';
 import { uniqBy } from 'lodash-es';
-import { boardsApi } from 'services/api/endpoints/boards';
-import { imagesApi } from 'services/api/endpoints/images';
 import type { ImageDTO } from 'services/api/types';
 
 import type { BoardId, ComparisonMode, GalleryState, GalleryView } from './types';
@@ -115,29 +113,6 @@ export const gallerySlice = createSlice({
       state.shouldShowArchivedBoards = action.payload;
     },
   },
-  extraReducers: (builder) => {
-    builder.addMatcher(isAnyBoardDeleted, (state, action) => {
-      const deletedBoardId = action.meta.arg.originalArgs;
-      if (deletedBoardId === state.selectedBoardId) {
-        state.selectedBoardId = 'none';
-        state.galleryView = 'images';
-      }
-      if (deletedBoardId === state.autoAddBoardId) {
-        state.autoAddBoardId = 'none';
-      }
-    });
-    builder.addMatcher(boardsApi.endpoints.listAllBoards.matchFulfilled, (state, action) => {
-      const boards = action.payload;
-
-      if (!state.autoAddBoardId) {
-        return;
-      }
-
-      if (!boards.map((b) => b.board_id).includes(state.autoAddBoardId)) {
-        state.autoAddBoardId = 'none';
-      }
-    });
-  },
 });
 
 export const {
@@ -161,11 +136,6 @@ export const {
   limitChanged,
   shouldShowArchivedBoardsChanged,
 } = gallerySlice.actions;
-
-const isAnyBoardDeleted = isAnyOf(
-  imagesApi.endpoints.deleteBoard.matchFulfilled,
-  imagesApi.endpoints.deleteBoardAndImages.matchFulfilled
-);
 
 export const selectGallerySlice = (state: RootState) => state.gallery;
